fix(mkBunny): enable zsh
This commit makes it so that mkBunny will also enable zsh. This silences a build error on systems that do not have zsh enabled. This commit has store path equality on Desktop and Laptop, **but introduces changes on Server.** However, I'm pushing this anyway, as the resulting changes should be of no consequence.